feat: game 우승자 계산해서 구하는 함수 util쪽으로 분리하여 구현
